+++ pkg/router/introspector.go
@@ -7,7 +7,11 @@ import (
 	"github.com/go-mojito/mojito/pkg/introspector"
 )
 
-var handlerIntrospector introspector.Introspector[HandlerIntrospectorArgFactory, builtinHandlerIntrospection]
+var (
+	errorInterface      = reflect.TypeOf((*error)(nil)).Elem()
+	handlerIntrospector introspector.Introspector[HandlerIntrospectorArgFactory, builtinHandlerIntrospection]
+	nextFuncTypeStdlib  = reflect.TypeOf((*http.Handler)(nil)).Elem()
+)
 
 func init() {
 	i, err := introspector.NewIntrospector[HandlerIntrospectorArgFactory, builtinHandlerIntrospection]()
@@ -50,6 +54,28 @@ type HandlerIntrospection interface {
 
 	// IsStdlibMiddleware returns true when the handler func is a stdlib middleware
 	IsStdlibMiddleware() bool
+}
+
+// HandlerIntrospection is a structure that contains all introspected details
+// about a (potential) handler
+type builtinHandlerIntrospection struct {
+	introspector.IntrospectorResult[HandlerIntrospectorArgFactory]
+}
+
+// CanError returns true when the handler func is returning error
+func (h builtinHandlerIntrospection) CanError() bool {
+	if h.Type().NumOut() != 1 {
+		return false
+	}
+	return h.Type().Out(0).Implements(errorInterface)
+}
+
+// IsStdlibMiddleware returns true when the handler func is a stdlib middleware
+func (h builtinHandlerIntrospection) IsStdlibMiddleware() bool {
+	if h.Type().NumOut() != 1 {
+		return false
+	}
+	return h.Type().Out(0).AssignableTo(nextFuncTypeStdlib)
 }
 
 // RegisterHandlerArgFactory will register a factory function for the given generic type
